+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 
 * In the ``preview_feed_list`` plugin, add ``<link rel=alternative ...>``
   tags as a feed detection heuristic.
+* In the web application, fix bug causing the entries page to crash
+  when counts are enabled.
 
 
 Version 1.12