+++ CHANGES.rst
@@ -6,6 +6,12 @@ Changelog
   :noindex:
 
 
+Version 1.14
+------------
+
+Unreleased
+
+
 Version 1.13
 ------------